Всё в одном файле :(
