--- MovieQuiz/IMDBGateway/IMDBMovie.swift
@@ -1,4 +1,0 @@
-struct IMDBMovie {
-    let name: String
-    let rating: Double
-}
