--- MovieQuiz/MovieQuizModel/Answer.swift
@@ -1,4 +1,0 @@
-enum Answer {
-    case yes
-    case no
-}
