--- MovieQuiz/MovieQuizModel/RiddleSign.swift
@@ -1,4 +1,0 @@
-enum RiddleSign: String {
-    case More = "больше"
-    case Less = "меньше"
-}
